rework resume view into experience and skills view
