+++ src/customizations/timeline.customization.tsx
@@ -3,7 +3,7 @@ import {
   TimelineConnector,
   TimelineContent,
   TimelineDot,
-  TimelineOppositeContent,
+  TimelineItem,
   TimelineSeparator,
 } from '@mui/lab';
 import { Theme } from '@mui/material';
@@ -11,21 +11,19 @@ import withStyles from '@mui/styles/withStyles';
 
 export const CustomTimeline = withStyles((theme: Theme) => ({}))(Timeline);
 
+export const CustomTimelineItem = withStyles(() => ({
+  missingOppositeContent: {
+    '&:before': {
+      display: 'none',
+    },
+  },
+}))(TimelineItem);
+
 export const CustomTimelineContent = withStyles((theme: Theme) => ({
   root: {
     paddingTop: 30,
   },
 }))(TimelineContent);
-
-export const CustomTimelineOppositeContent = withStyles((theme: Theme) => ({
-  root: {
-    fontWeight: 'bold',
-    paddingTop: 40,
-    fontSize: 'calc(1vh + 1.2vw)',
-    display: 'flex',
-    justifyContent: 'center',
-  },
-}))(TimelineOppositeContent);
 
 export const CustomTimelineSeparator = withStyles((theme: Theme) => ({
   root: {},
